Changed all .js to .jsx
