+++ src/components/PopUpRegistration.jsx
